I've slightly refactored the code to make it much more readable and to add more documentation about what each bit does (so I won't forget when I come back to read this again later!)
[ci skip]
